requirements.txt
plotly==5.2.1
streamlit==0.86.0
pandas==1.3.2
Pillow==8.3.1
numpy==1.21.2
openpyxl==3.0.7